+++ operation/companies/company_role.php
@@ -24,7 +24,11 @@ if (! give_acl ($config["id_user"], 0, "VR")) {
 	exit;
 }
 
+$manager = give_acl ($config["id_user"], 0, "VM");
+
 $id = (int) get_parameter ('id');
+$company = get_db_row ('tcompany', 'id', $id);
+$id_group = $company['id_grupo'];	
 $new_role = (bool) get_parameter ('new_role');
 $create_role = (bool) get_parameter ('create_role');
 $update_role = (bool) get_parameter ('update_role');
@@ -54,12 +58,11 @@ if ($create_role) {
 
 // UPDATE
 if ($update_role) {
-
-	if (! give_acl ($config["id_user"], 0, "VM")) {
+	if (! give_acl ($config["id_user"], $id_group, "VW")) {
                audit_db ($config["id_user"], $config["REMOTE_ADDR"], "ACL Violation","Trying to update a company role");
                require ("general/noaccess.php");
                exit;
-        }
+    }
 
 	$name = (string) get_parameter ('name');
 	$description = (string) get_parameter ('description');
@@ -80,11 +83,11 @@ if ($update_role) {
 
 // DELETE
 if ($delete_role) {
-        if (! give_acl ($config["id_user"], 0, "VM")) {
-                audit_db ($config["id_user"], $config["REMOTE_ADDR"], "ACL Violation","Trying to delete a company role");
-                require ("general/noaccess.php");
-                exit;
-        }
+	if (! give_acl ($config["id_user"], $id_group, "VM")) {
+			audit_db ($config["id_user"], $config["REMOTE_ADDR"], "ACL Violation","Trying to delete a company role");
+			require ("general/noaccess.php");
+			exit;
+	}
 
 	$name = get_db_value ('name', 'tcompany_role', 'id', $id);
 	$sql = sprintf ('DELETE FROM tcompany_role WHERE id = %d', $id);
@@ -99,9 +102,19 @@ echo "<h2>".__('Company role management')."</h2>";
 // FORM (Update / Create)
 if ($id || $new_role) {
 	if ($new_role) {
+		if(!$manager) {
+			audit_db ($config["id_user"], $config["REMOTE_ADDR"], "ACL Violation","Trying to add a company role");
+			require ("general/noaccess.php");
+			exit;
+		}
 		$name = '';
 		$description = '';
 	} else {
+		if (! give_acl ($config["id_user"], $id_group, "VR")) {
+				   audit_db ($config["id_user"], $config["REMOTE_ADDR"], "ACL Violation","Trying to update a company role");
+				   require ("general/noaccess.php");
+				   exit;
+		}
 		$role = get_db_row ('tcompany_role', 'id', $id);
 		$name = $role['name'];
 		$description = $role['description'];
@@ -112,21 +125,32 @@ if ($id || $new_role) {
 	$table->data = array ();
 	$table->colspan = array ();
 	
-	$table->data[0][0] = print_input_text ("name", $name, "", 60, 100, true, __('Role name'));
-	$table->data[1][0] = print_textarea ('description', 14, 1, $description, '', true, __('Description'));
+	if (give_acl ($config["id_user"], $id_group, "VW")) {
+		$table->data[0][0] = print_input_text ("name", $name, "", 60, 100, true, __('Role name'));
+		$table->data[1][0] = print_textarea ('description', 14, 1, $description, '', true, __('Description'));
+	}
+	else {
+		$table->data[0][0] = "<b>".__('Role name')."</b><br>$name<br>";
+		if($description == '') {
+			$description = '<i>-'.__('Empty').'-</i>';
+		}		
+		$table->data[1][0] = "<b>".__('Description')."</b><br>$description<br>";
+	}
 	
 	echo '<form method="post" action="index.php?sec=customers&sec2=operation/companies/company_role">';
-	print_table ($table);
-	echo '<div class="button" style="width: '.$table->width.'">';
-	if ($id) {
-		print_submit_button (__('Update'), "update_btn", false, 'class="sub upd"', false);
-		print_input_hidden ('update_role', 1);
-		print_input_hidden ('id', $id);
-	} else {
-		print_input_hidden ('create_role', 1);
-		print_submit_button (__('Create'), "create_btn", false, 'class="sub next"', false);
-	}
-	echo "</div>";
+		print_table ($table);
+		if (($id && give_acl ($config["id_user"], $id_group, "VW")) || (!$id && give_acl ($config["id_user"], $id_group, "VM"))) {
+			echo '<div class="button" style="width: '.$table->width.'">';
+			if ($id) {
+				print_submit_button (__('Update'), "update_btn", false, 'class="sub upd"', false);
+				print_input_hidden ('update_role', 1);
+				print_input_hidden ('id', $id);
+			} else {
+				print_input_hidden ('create_role', 1);
+				print_submit_button (__('Create'), "create_btn", false, 'class="sub next"', false);
+			}
+			echo "</div>";
+		}
 	echo '</form>';
 } else {
 	$search_text = (string) get_parameter ('search_text');
@@ -163,7 +187,9 @@ if ($id || $new_role) {
 		$table->style[0] = 'font-weight: bold';
 		$table->head[0] = __('Name');
 		$table->head[1] = __('Description');
-		$table->head[2] = __('Delete');
+		if(give_acl ($config["id_user"], $id_group, "VM")) {
+			$table->head[2] = __('Delete');
+		}
 		
 		foreach ($roles as $role) {
 			$data = array ();
@@ -171,21 +197,26 @@ if ($id || $new_role) {
 			$data[0] = "<a href='index.php?sec=customers&sec2=operation/companies/company_role&id=".
 				$role["id"]."'>".$role["name"]."</a>";
 			$data[1] = substr ($role["description"], 0, 70)."...";
-			$data[3] = '<a href="index.php?sec=customers&
-						sec2=operation/companies/company_role&
-						delete_role=1&id='.$role['id'].'"
-						onClick="if (!confirm(\''.__('Are you sure?').'\'))
-						return false;">
-						<img src="images/cross.png"></a>';
+			if(give_acl ($config["id_user"], $id_group, "VM")) {
+				$data[2] = '<a href="index.php?sec=customers&
+							sec2=operation/companies/company_role&
+							delete_role=1&id='.$role['id'].'"
+							onClick="if (!confirm(\''.__('Are you sure?').'\'))
+							return false;">
+							<img src="images/cross.png"></a>';
+			}
 			array_push ($table->data, $data);
 		}
 		print_table ($table);
 	}
 	
-	echo '<form method="post" action="index.php?sec=customers&sec2=operation/companies/company_role">';
-	echo '<div class="button" style="width: '.$table->width.'">';
-	print_submit_button (__('Create'), 'new_btn', false, 'class="sub next"');
-	print_input_hidden ('new_role', 1);
-	echo '</div>';
+	if($manager) {
+		echo '<form method="post" action="index.php?sec=customers&sec2=operation/companies/company_role">';
+		echo '<div class="button" style="width: '.$table->width.'">';
+		print_submit_button (__('Create'), 'new_btn', false, 'class="sub next"');
+		print_input_hidden ('new_role', 1);
+		echo '</div>';
+		echo '</form>';
+	}
 }
 ?>
